+++ llvm-parser/plugins/de.upb.llvm_parser/src-gen/de/upb/llvm_parser/llvm/validation/StoreValidator.java
@@ -8,6 +8,8 @@ import de.upb.llvm_parser.llvm.TypeList;
 import de.upb.llvm_parser.llvm.TypeUse;
 import de.upb.llvm_parser.llvm.Value;
 
+import java.math.BigDecimal;
+
 import org.eclipse.emf.common.util.EList;
 
 /**
@@ -25,5 +27,5 @@ public interface StoreValidator {
 	boolean validateAdresstype(TypeUse value);
 	boolean validateAdress(Value value);
 	boolean validateOrdering(String value);
-	boolean validateIndex(int value);
+	boolean validateIndex(BigDecimal value);
 }
